Split out builders from sold.cc
