+++ sold.cc
@@ -26,188 +26,9 @@
 #include "elf_binary.h"
 #include "hash.h"
 #include "ldsoconf.h"
+#include "strtab_builder.h"
+#include "symtab_builder.h"
 #include "utils.h"
-
-class StrtabBuilder {
-public:
-    uintptr_t Add(const std::string& s) {
-        CHECK(!is_freezed_);
-        uintptr_t pos = static_cast<uintptr_t>(strtab_.size());
-        strtab_ += s;
-        strtab_ += '\0';
-        return pos;
-    }
-
-    void Freeze() { is_freezed_ = true; }
-
-    size_t size() const { return strtab_.size(); }
-
-    const void* data() const { return strtab_.data(); }
-
-private:
-    std::string strtab_;
-    bool is_freezed_{false};
-};
-
-class SymtabBuilder {
-public:
-    SymtabBuilder() {
-        Symbol sym = {0};
-        AddSym("");
-        CHECK(syms_.emplace("", sym).second);
-    }
-
-    struct Symbol {
-        Elf_Sym sym;
-        uintptr_t index;
-    };
-
-    void SetSrcSyms(std::map<std::string, Elf_Sym*> syms) { src_syms_ = syms; }
-
-    uintptr_t AddSym(const std::string& name) {
-        uintptr_t index = sym_names_.size();
-        sym_names_.push_back(name);
-        return index;
-    }
-
-    bool Resolve(const std::string& name, uintptr_t* val_or_index) {
-        Symbol sym{};
-        sym.sym.st_name = 0;
-        sym.sym.st_info = 0;
-        sym.sym.st_other = 0;
-        sym.sym.st_shndx = 0;
-        sym.sym.st_value = 0;
-        sym.sym.st_size = 0;
-
-        auto found = syms_.find(name);
-        if (found != syms_.end()) {
-            sym = found->second;
-        } else {
-            auto found = src_syms_.find(name);
-            if (found != src_syms_.end()) {
-                sym.sym = *found->second;
-                if (IsDefined(sym.sym)) {
-                    LOGF("Symbol %s found\n", name.c_str());
-                } else {
-                    LOGF("Symbol (undef/weak) %s found\n", name.c_str());
-                    sym.index = AddSym(name);
-                    CHECK(syms_.emplace(name, sym).second);
-                }
-            } else {
-                LOGF("Symbol %s not found\n", name.c_str());
-                sym.index = AddSym(name);
-                CHECK(syms_.emplace(name, sym).second);
-            }
-        }
-
-        if (!sym.sym.st_value) {
-            *val_or_index = sym.index;
-            return false;
-        } else {
-            *val_or_index = sym.sym.st_value;
-            return true;
-        }
-    }
-
-    uintptr_t ResolveCopy(const std::string& name) {
-        // TODO(hamaji): Refactor.
-        Symbol sym{};
-        sym.sym.st_name = 0;
-        sym.sym.st_info = 0;
-        sym.sym.st_other = 0;
-        sym.sym.st_shndx = 0;
-        sym.sym.st_value = 0;
-        sym.sym.st_size = 0;
-
-        auto found = syms_.find(name);
-        if (found != syms_.end()) {
-            sym = found->second;
-        } else {
-            auto found = src_syms_.find(name);
-            if (found != src_syms_.end()) {
-                LOGF("Symbol %s found for copy\n", name.c_str());
-                sym.sym = *found->second;
-                sym.index = AddSym(name);
-                CHECK(syms_.emplace(name, sym).second);
-            } else {
-                LOGF("Symbol %s not found for copy\n", name.c_str());
-                CHECK(false);
-            }
-        }
-
-        return sym.index;
-    }
-
-    void Build(StrtabBuilder* strtab) {
-        for (const std::string& name : sym_names_) {
-            auto found = syms_.find(name);
-            CHECK(found != syms_.end());
-            Elf_Sym sym = found->second.sym;
-            sym.st_name = strtab->Add(name);
-            symtab_.push_back(sym);
-        }
-    }
-
-    void AddPublicSymbol(const std::string& name, Elf_Sym sym) { public_syms_.emplace(name, sym); }
-
-    void MergePublicSymbols(StrtabBuilder* strtab) {
-        gnu_hash_.nbuckets = 1;
-        CHECK(symtab_.size() <= std::numeric_limits<uint32_t>::max());
-        gnu_hash_.symndx = symtab_.size();
-        gnu_hash_.maskwords = 1;
-        gnu_hash_.shift2 = 1;
-
-        for (const auto& p : public_syms_) {
-            const std::string& name = p.first;
-            Elf_Sym sym = p.second;
-            sym.st_name = strtab->Add(name);
-            sym.st_shndx = 1;
-            sym_names_.push_back(name);
-            symtab_.push_back(sym);
-        }
-        public_syms_.clear();
-    }
-
-    uintptr_t size() const { return symtab_.size() + public_syms_.size(); }
-
-    const Elf_GnuHash& gnu_hash() const { return gnu_hash_; }
-
-    uintptr_t gnu_hash_size() const {
-        CHECK(!symtab_.empty());
-        CHECK(public_syms_.empty());
-        CHECK(gnu_hash_.nbuckets);
-        return (sizeof(uint32_t) * 4 + sizeof(Elf_Addr) + sizeof(uint32_t) * (1 + symtab_.size() - gnu_hash_.symndx));
-    }
-
-    const std::vector<Elf_Sym>& Get() { return symtab_; }
-
-    const std::vector<std::string>& GetNames() const { return sym_names_; }
-
-private:
-    std::map<std::string, Elf_Sym*> src_syms_;
-    std::map<std::string, Symbol> syms_;
-
-    std::vector<std::string> sym_names_;
-    std::vector<Elf_Sym> symtab_;
-    std::map<std::string, Elf_Sym> public_syms_;
-
-    Elf_GnuHash gnu_hash_;
-};
-
-struct TLS {
-    struct Data {
-        ELFBinary* bin;
-        uint8_t* start;
-        size_t size;
-        uintptr_t file_offset;
-        uintptr_t bss_offset;
-    };
-
-    std::vector<Data> data;
-    std::map<ELFBinary*, size_t> bin_to_index;
-    uintptr_t filesz{0};
-    uintptr_t memsz{0};
-};
 
 class Sold {
 public:
